requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
python-dotenv>=1.0
slowapi>=0.1.9
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
psycopg2-binary>=2.9.11 
sqlalchemy>=2.0.48